Introducing separate nlo-repo script
Within NLO, libraries with their own repository in general processed by the script nlo-repo.py, situated at the root of the repository. This script describes all libraries, accessible in this repository, in this particular case - it is single library 'zoo'.

Before this change LensCore was emulating this service for the zoo library, but that creates issues with export_sources_git attribute, so now we need pedantically correct layout.